refactor(file): 检查文件是否存在的功能应该放到 file.js 里
