+++ utils/file.js
@@ -1,4 +1,5 @@
 import fs from "node:fs";
+import util from "util";
 
 class MucFile{
     create(path, data){
@@ -9,6 +10,15 @@ class MucFile{
                 console.info("\n文件创建成功！\n" + data)
             }
         })
+    }
+
+    async fileExistCheck(name) {
+        try {
+            let stat = await util.promisify(fs.stat)(name);
+            return stat.isFile() === true
+        } catch (err) {
+            return false
+        }
     }
 }
